+++ app/assets/javascripts/govuk_publishing_components/components/add-another.js
@@ -1,3 +1,4 @@
+//= require sortablejs/Sortable.js
 window.GOVUK = window.GOVUK || {}
 window.GOVUK.Modules = window.GOVUK.Modules || {};
 
@@ -27,6 +28,9 @@ window.GOVUK.Modules = window.GOVUK.Modules || {};
     this.createAddAnotherButton()
     this.createRemoveButtons()
     this.removeEmptyFieldset()
+    if (this.module.dataset.sortable === 'true') {
+      this.makeSortable()
+    }
     this.updateFieldsetsAndButtons()
   }
 
@@ -102,12 +106,14 @@ window.GOVUK.Modules = window.GOVUK.Modules || {};
     visibleFields.forEach(function (field, index) {
       field.querySelector('legend').textContent = this.module.dataset.fieldsetLegend + ' ' + (index + 1)
 
-      var trackedRemoveButton = field.parentNode.querySelector('.js-add-another__remove-button[data-ga4-event]')
+      var trackedButtons = field.parentNode.querySelectorAll(
+        '.js-add-another__remove-button[data-ga4-event], .gem-c-add-another__move-button[data-ga4-event]'
+      )
 
-      if (trackedRemoveButton) {
-        trackedRemoveButton.dataset.indexSection = this.startIndex + index
-        trackedRemoveButton.dataset.indexSectionCount = this.indexSectionCount || visibleFields.length
-      }
+      trackedButtons.forEach(function (button) {
+        button.dataset.indexSection = this.startIndex + index
+        button.dataset.indexSectionCount = this.indexSectionCount || visibleFields.length
+      }.bind(this))
     }.bind(this))
 
     if (this.module.dataset.emptyFields === 'false') {
@@ -115,6 +121,10 @@ window.GOVUK.Modules = window.GOVUK.Modules || {};
         'js-add-another__remove-button--hidden',
         this.module.querySelectorAll('.js-add-another__fieldset:not([hidden])').length === 1
       )
+    }
+
+    if (this.module.dataset.sortable === 'true') {
+      this.disableTopAndBottomButtons()
     }
 
     var trackedAddAnotherButton = this.module.querySelector('.js-add-another__add-button[data-ga4-event]')
@@ -125,15 +135,30 @@ window.GOVUK.Modules = window.GOVUK.Modules || {};
     }
   }
 
+  // Disable the top "Move up" and bottom "Move down" buttons when the list is sortable.
+  AddAnother.prototype.disableTopAndBottomButtons = function () {
+    this.module.querySelectorAll('.gem-c-add-another__move-button').forEach(function (button) {
+      button.removeAttribute('disabled')
+    })
+
+    var topMoveUpButton = this.module.querySelector('.js-add-another__fieldset:first-of-type .gem-c-add-another__move-button[data-action="move-up"]')
+    topMoveUpButton.setAttribute('disabled', true)
+
+    var bottomMoveDownButton = this.module.querySelector('.js-add-another__fieldset:last-of-type .gem-c-add-another__move-button[data-action="move-down"]')
+    bottomMoveDownButton.setAttribute('disabled', true)
+  }
+
   AddAnother.prototype.addNewFieldset = function (event) {
-    var button = event.target
     var newFieldsetTemplate = this.module.querySelector('.js-add-another__empty-template')
     var newFieldset = newFieldsetTemplate.content.cloneNode(true).children[0]
     newFieldset.classList.add('js-add-another__fieldset')
     this.createRemoveButton(newFieldset, this.removeNewFieldset.bind(this))
-    button.before(newFieldset)
+    newFieldsetTemplate.before(newFieldset)
 
-    this.incrementAttributes(newFieldsetTemplate.content)
+    this.updateAttributes(newFieldsetTemplate.content)
+    if (this.module.dataset.sortable === 'true') {
+      this.makeSortable()
+    }
     this.updateFieldsetsAndButtons()
 
     // Initialise any modules included in new fieldset
@@ -160,7 +185,7 @@ window.GOVUK.Modules = window.GOVUK.Modules || {};
   }
 
   // Set attribute values for id, for and name of supplied fieldset
-  AddAnother.prototype.incrementAttributes = function (fieldset) {
+  AddAnother.prototype.updateAttributes = function (fieldset, newIndex = null) {
     var matcher = /(.*[_[])([0-9]+)([_\]].*?)$/
     fieldset
       .querySelectorAll('label, input, select, textarea')
@@ -169,10 +194,116 @@ window.GOVUK.Modules = window.GOVUK.Modules || {};
           var value = element.getAttribute(attribute)
           var matched = matcher.exec(value)
           if (!matched) return
-          var index = parseInt(matched[2], 10) + 1
+          var index = newIndex == null ? (parseInt(matched[2], 10) + 1) : newIndex
           element.setAttribute(attribute, matched[1] + index + matched[3])
         })
       })
+  }
+
+  AddAnother.prototype.makeSortable = function () {
+    this.removeMoveButtons()
+    this.addMoveButtons()
+    this.hideOrderInputs()
+    this.sortableList = window.Sortable.create(this.module, {
+      dataIdAttr: 'data-sortable-id',
+      disabled: true
+    })
+  }
+
+  AddAnother.prototype.getOrderInputs = function () {
+    return this.module.querySelectorAll('.js-add-another__fieldset:not([hidden]) .js-add-another__order-input input')
+  }
+
+  AddAnother.prototype.hideOrderInputs = function () {
+    this.getOrderInputs().forEach(function (orderInput) {
+      orderInput.closest('.js-add-another__order-input').hidden = true
+    })
+  }
+
+  AddAnother.prototype.addMoveButtons = function () {
+    var visibleFields = this.module.querySelectorAll('.js-add-another__fieldset:not([hidden])')
+    visibleFields.forEach(function (field, index) {
+      field.dataset.sortableId = 'js-add-another__sortable-' + index + '-fieldset'
+      var wrapper = document.createElement('div')
+      wrapper.className = 'gem-c-add-another__move-button-wrapper'
+      this.addButton(wrapper, 'move-up')
+      this.addButton(wrapper, 'move-down')
+      field.prepend(wrapper)
+    }.bind(this))
+  }
+
+  AddAnother.prototype.addButton = function (wrapper, action) {
+    var button = document.createElement('button')
+    button.className = 'govuk-button gem-c-button--secondary-quiet gem-c-add-another__move-button'
+    button.innerText = action === 'move-up' ? 'Move up' : 'Move down'
+    button.dataset.action = action
+    if (!this.disableGa4) {
+      button.dataset.ga4Event = this.createEventData({ action: action })
+    }
+
+    button.addEventListener('click', this.moveFieldset.bind(this))
+    wrapper.prepend(button)
+  }
+
+  AddAnother.prototype.moveFieldset = function (event) {
+    event.preventDefault()
+
+    var button = event.currentTarget
+    var action = button.dataset.action
+    var fieldset = button.closest('.js-add-another__fieldset')
+
+    var sortableId = fieldset.dataset.sortableId
+
+    var order = this.sortableList.toArray()
+    var index = order.indexOf(sortableId)
+
+    // pull the item we're moving out of the order
+    order.splice(index, 1)
+
+    // put it back in at the correct position
+    if (action === 'move-down') {
+      order.splice(index + 1, 0, sortableId)
+    } else if (action === 'move-up') {
+      order.splice(index - 1, 0, sortableId)
+    }
+
+    this.sortableList.sort(order, true)
+    this.updateFieldsetsAndButtons()
+
+    var visibleFields = this.module.querySelectorAll('.js-add-another__fieldset:not([hidden])')
+    visibleFields.forEach(function (field, index) {
+      this.updateAttributes(field, index)
+    }.bind(this))
+
+    var triggeredByKeyboard = event.detail === 0
+    if (triggeredByKeyboard) {
+      this.preserveElementFocus(button, fieldset, action, visibleFields)
+    }
+
+    this.updateOrderInputs()
+  }
+
+  AddAnother.prototype.preserveElementFocus = function (button, fieldset, action, visibleFields) {
+    let target = button
+    if (action === 'move-up' && fieldset === visibleFields[0]) {
+      target = button.parentNode.querySelector('button[data-action="move-down"]')
+    } else if (action === 'move-down' && fieldset === visibleFields[visibleFields.length - 1]) {
+      target = button.parentNode.querySelector('button[data-action="move-up"]')
+    }
+    target.focus()
+  }
+
+  AddAnother.prototype.removeMoveButtons = function () {
+    var wrappers = this.module.querySelectorAll('.gem-c-add-another__move-button-wrapper')
+    wrappers.forEach((button) => {
+      button.remove()
+    })
+  }
+
+  AddAnother.prototype.updateOrderInputs = function () {
+    this.getOrderInputs().forEach(function (orderInput, index) {
+      orderInput.value = index + 1
+    })
   }
 
   Modules.AddAnother = AddAnother