+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Adjust pagination component ([PR #5176](https://github.com/alphagov/govuk_publishing_components/pull/5176))
 * Guess `world_location_news` paths by `document_type` ([PR #5177](https://github.com/alphagov/govuk_publishing_components/pull/5177))
 * Add image card helper spec and refactor ([PR #5179](https://github.com/alphagov/govuk_publishing_components/pull/5179))
+* Allow add another fieldsets to be reordered ([PR #5170](https://github.com/alphagov/govuk_publishing_components/pull/5170))
 
 ## 63.0.0